backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "semiovis-backend"
version = "1.0.0"
description = "SemioVis — Semiotic image analysis backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=22.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "Pillow>=10.0",
    "numpy>=1.26",
    "opencv-python>=4.8",
    "onnxruntime>=1.17",
    "ultralytics>=8.2",
    "mediapipe>=0.10",
    "reportlab>=4.0",
    "python-docx>=1.0",
    "Jinja2>=3.1",
]

[project.optional-dependencies]
saas = [
    "celery>=5.3",
    "redis>=5.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../tests/backend"]
